slippage redesign for composite swap screen
