+++ mobile-app/app/screens/AppNavigator/screens/Dex/CompositeSwap/components/SlippageTolerance.tsx
@@ -1,124 +1,211 @@
-import { useCallback, useEffect, useState } from 'react'
+import { ReactElement, useEffect, useState } from 'react'
 import { TouchableOpacity } from 'react-native'
 import BigNumber from 'bignumber.js'
-import NumberFormat from 'react-number-format'
 import { tailwind } from '@tailwind'
 import { translate } from '@translations'
-import { useBottomSheetModal } from '@gorhom/bottom-sheet'
 import { ThemedIcon, ThemedText, ThemedView } from '@components/themed'
 import { View } from '@components'
-import { BottomSheetSlippageTolerance } from './BottomSheetSlippageTolerance'
-import { BottomSheetNavScreen } from '@components/BottomSheetWithNav'
+import { BottomSheetInfo } from '@components/BottomSheetInfo'
+import { WalletTextInput } from '@components/WalletTextInput'
+import { debounce } from 'lodash'
 
-const SLIPPAGE_MODAL_NAME = 'SlippageTolerance'
 const percentageList = ['1', '3', '5', '10', '20']
 
 interface SlippageToleranceProps {
   slippage: BigNumber
+  slippageError?: SlippageError
+  setSlippageError: (error?: SlippageError) => void
   setSlippage: (slippage: BigNumber) => void
-  setIsSelectorOpen: (val: boolean) => void
-  setBottomSheetScreen: (screens: BottomSheetNavScreen[]) => void
-  onPress: () => void
-  onCloseButtonPress: () => void
 }
 
-export function SlippageTolerance ({ slippage, setSlippage, setIsSelectorOpen, setBottomSheetScreen, onPress, onCloseButtonPress }: SlippageToleranceProps): JSX.Element {
-  const { dismiss } = useBottomSheetModal()
+export function SlippageTolerance ({ slippage, setSlippage, slippageError, setSlippageError }: SlippageToleranceProps): JSX.Element {
   const [isCustomSlippage, setIsCustomSlippage] = useState(false)
-  const onSubmitSlippage = (val: BigNumber, isCustom: boolean): void => {
-    setIsCustomSlippage(isCustom)
-    setSlippage(val)
-    setIsSelectorOpen(false)
-    dismiss(SLIPPAGE_MODAL_NAME)
+  const slippageTolerance = {
+    title: 'Slippage tolerance',
+    message: 'Slippages are rate changes that occur within an order transaction. Choose how much of this slippage are you willing to accept.'
   }
-
-  const setSlippageToleranceToBottomSheet = useCallback(() => {
-    setBottomSheetScreen([
-      {
-        stackScreenName: 'BottomSheetSlippageTolerance',
-        component: BottomSheetSlippageTolerance({
-          slippage,
-          isCustomSlippage,
-          onSubmitSlippage,
-          onCloseButtonPress
-        }),
-        option: {
-          header: () => null
-        }
-      }])
-  }, [slippage, isCustomSlippage])
 
   useEffect(() => {
     setIsCustomSlippage(!percentageList.includes(new BigNumber(slippage).toString()))
-  }, [slippage])
+  }, [])
 
   return (
     <ThemedView
-      dark={tailwind('bg-gray-800 border-b border-gray-700')}
-      light={tailwind('bg-white border-b border-gray-200')}
       style={tailwind('flex-row items-center w-full')}
     >
-      <View style={tailwind('w-full')}>
-        <TouchableOpacity
-          onPress={() => {
-            setIsSelectorOpen(true)
-            setSlippageToleranceToBottomSheet()
-            onPress()
-          }}
-          style={tailwind('p-4 pr-2')}
-          testID='slippage_select'
+      <View style={tailwind('px-4 mt-5 w-full')}>
+        <View
+          testID='slippage_tolerance_heading'
+          style={tailwind(['flex-row mb-3 items-center justify-between'])}
         >
-          <SlippageToleranceRow slippage={slippage} />
-        </TouchableOpacity>
+          <View style={tailwind('flex-row items-center')}>
+            <ThemedText
+              dark={tailwind('text-gray-50')}
+              light={tailwind('text-gray-900')}
+              style={tailwind('mr-1')}
+            >
+              {translate('screens/SlippageTolerance', 'Slippage tolerance')}
+            </ThemedText>
+            <BottomSheetInfo alertInfo={slippageTolerance} name={slippageTolerance.title} />
+          </View>
+        </View>
+        <SlippageSelector
+          slippage={slippage}
+          slippageError={slippageError}
+          setSlippageError={setSlippageError}
+          isCustomSlippage={isCustomSlippage}
+          setIsCustomSlippage={setIsCustomSlippage}
+          onSubmitSlippage={setSlippage}
+        />
       </View>
     </ThemedView>
   )
 }
 
-function SlippageToleranceRow ({ slippage }: { slippage: BigNumber}): JSX.Element {
+interface SlippageSelectorProps {
+  isCustomSlippage: boolean
+  onSubmitSlippage: (val: BigNumber, isCustomSlippage: boolean) => void
+  setIsCustomSlippage: (val: boolean) => void
+  slippage: BigNumber
+  slippageError?: SlippageError
+  setSlippageError: (error?: SlippageError) => void
+}
+
+export interface SlippageError {
+  type: 'error' | 'helper'
+  text?: string
+}
+
+function SlippageSelector ({ isCustomSlippage, onSubmitSlippage, slippage, setIsCustomSlippage, slippageError, setSlippageError }: SlippageSelectorProps): JSX.Element {
+  const percentageList = ['1', '3', '5', '10', '20']
+  const [selectedSlippage, setSelectedSlippage] = useState(slippage.toString())
+  const [isRiskWarningDisplayed, setIsRiskWarningDisplayed] = useState(false)
+  const submitSlippage = debounce(onSubmitSlippage, 500)
+
+  const onSlippageChange = (value: string): void => {
+    setSelectedSlippage(value)
+    submitSlippage(new BigNumber(value), isCustomSlippage)
+  }
+
+  const isSlippageValid = (): boolean => {
+    return slippageError === undefined || slippageError?.type === 'helper'
+  }
+
+  const validateSlippage = (value: string): void => {
+    if (value === undefined || value === '') {
+      setSlippageError({
+        type: 'error',
+        text: translate('screens/SlippageTolerance', 'Required field is missing')
+      })
+      return
+    } else if (!(new BigNumber(value).gte(0) && new BigNumber(value).lte(100))) {
+      setSlippageError({
+        type: 'error',
+        text: translate('screens/SlippageTolerance', 'Slippage rate must range from 0-100%')
+      })
+      return
+    }
+
+    setSlippageError(undefined)
+  }
+
+  useEffect(() => {
+    validateSlippage(selectedSlippage)
+    setIsRiskWarningDisplayed((new BigNumber(selectedSlippage).gte(20) && new BigNumber(selectedSlippage).lte(100)))
+  }, [selectedSlippage])
+
   return (
-    <View
-      style={tailwind('flex-row')}
-    >
-      <View style={tailwind('w-6/12 flex-grow')}>
-        <ThemedText
-          dark={tailwind('text-gray-200')}
-          light={tailwind('text-black')}
-          style={tailwind('text-sm mr-1')}
-          testID='slippage_title'
-        >
-          {translate('screens/SlippageTolerance', 'Slippage tolerance')}
-        </ThemedText>
-      </View>
-      <View
-        style={tailwind('flex w-6/12 flex-row justify-end flex-wrap items-center')}
-      >
-        <NumberFormat
-          decimalScale={8}
-          displayType='text'
-          renderText={(val: string) => (
-            <>
-              <ThemedText
-                dark={tailwind('text-gray-400')}
-                light={tailwind('text-gray-500')}
-                style={tailwind('text-sm text-right')}
-                testID='slippage_value'
-              >
-                {val}%
-              </ThemedText>
-              <ThemedIcon
-                light={tailwind('text-primary-500')}
-                dark={tailwind('text-darkprimary-500')}
-                iconType='MaterialIcons'
-                name='unfold-more'
-                size={24}
-              />
-            </>
-        )}
-          thousandSeparator
-          value={new BigNumber(slippage).toNumber()}
+    <>
+      <View style={tailwind('flex-row mb-3 flex-wrap')}>
+        {percentageList.map((value) => (
+          <SlippageButton
+            key={value}
+            onPress={() => {
+              setIsCustomSlippage(false)
+              onSlippageChange(value)
+            }}
+            isActive={!isCustomSlippage && selectedSlippage === value}
+            label={`${value}%`}
+          />
+        ))}
+        <SlippageButton
+          onPress={() => {
+            setIsCustomSlippage(true)
+          }}
+          icon={
+            <ThemedIcon
+              size={14}
+              name='edit'
+              iconType='MaterialIcons'
+              style={tailwind('mr-1')}
+              dark={tailwind(`${isCustomSlippage ? 'text-darkprimary-500' : ''}`)}
+              light={tailwind(`${isCustomSlippage ? 'text-primary-500' : ''}`)}
+            />
+          }
+          isActive={isCustomSlippage}
+          label={translate('screens/SlippageTolerance', 'Custom')}
         />
       </View>
+
+      {isCustomSlippage && (
+        <ThemedView
+          dark={tailwind('bg-transparent')}
+          light={tailwind('bg-transparent')}
+          style={tailwind('flex-row w-full')}
+        >
+          <WalletTextInput
+            autoFocus
+            onChangeText={onSlippageChange}
+            keyboardType='numeric'
+            autoCapitalize='none'
+            placeholder='0.00%'
+            style={tailwind('flex-grow h-8')}
+            containerStyle='mb-1 w-full flex-col'
+            testID='slippage_input'
+            value={selectedSlippage !== undefined ? selectedSlippage.toString() : ''}
+            displayClearButton={selectedSlippage !== undefined}
+            onClearButtonPress={() => onSlippageChange('')}
+            inputType='numeric'
+            valid={isSlippageValid()}
+            inlineText={slippageError}
+          />
+        </ThemedView>
+      )}
+
+      {isRiskWarningDisplayed && (
+        <ThemedText
+          light={tailwind('text-error-500')}
+          dark={tailwind('text-darkerror-500')}
+          style={tailwind('text-sm mb-3')}
+          testID='slippage_warning'
+        >
+          {translate('screens/SlippageTolerance', 'Proceed at your own risk')}
+        </ThemedText>
+      )}
+    </>
+  )
+}
+
+function SlippageButton ({ onPress, isActive, label, icon }: { onPress: () => void, isActive: boolean, label: string, icon?: ReactElement }): JSX.Element {
+  const buttonStyles = 'flex flex-row px-2 py-1.5 border border-gray-300 rounded items-center'
+  const activeStyle = 'bg-primary-50 border-primary-50 text-primary-500'
+  return (
+    <View style={tailwind('mr-1.5 mt-2')}>
+      <TouchableOpacity
+        key={label}
+        onPress={onPress}
+        style={tailwind(`${buttonStyles} ${isActive ? activeStyle : ''} `)}
+        testID={`slippage_${label}`}
+      >
+        {icon}
+        <ThemedText
+          dark={tailwind(`${isActive ? 'text-darkprimary-500 font-semibold' : ''}`)}
+          light={tailwind(`${isActive ? 'text-primary-500 font-semibold' : ''}`)}
+          style={tailwind('text-sm')}
+        >
+          {label}
+        </ThemedText>
+      </TouchableOpacity>
     </View>
   )
 }
